now checks first if ..\QJediInstaller.pas exists before it tries to rename it.

+++ jcl/install/prototypes/jpp.sh
@@ -10,4 +10,6 @@ OPTIONS="-c -dVisualCLX -dHAS_UNIT_TYPES -uDevelop -uVCL -x../Q"
 FILES="JclInstall.pas JediInstallIntf.pas ProductFrames.pas JediInstallerMain.pas JediInstaller.dpr"
 
 $JPP $OPTIONS $FILES
-mv ../QJediInstaller.pas ../QJediInstaller.dpr
+if [ -f ../QJediInstaller.pas ]; then
+  mv ../QJediInstaller.pas ../QJediInstaller.dpr
+fi
